feat(test): update tests to work on cluster with deployed apps and the k6 operator
