+++ README.md
@@ -36,9 +36,64 @@ scp <user>@<NODE_IP>:$HOME/.kube/config $HOME/.kube/config
 sed -i '' 's/0\.0\.0\.0/<NODE_IP>/g' $HOME/.kube/config
 ```
 
-## Executing a Test
+## Executing a Test with the k6 Operator
 
-Use the [`run.sh`](./tests/run.sh) to execute a test. For now, the script assumes a single node cluster. Specify the node IP address in `NODE_IP`.
+1. [Install the operator](https://github.com/grafana/k6-operator/tree/main?tab=readme-ov-file#deployment-with-helm)
+
+   ```sh
+   helm repo add grafana https://grafana.github.io/helm-charts
+    helm repo update
+    helm install k6-operator grafana/k6-operator
+    ```
+
+2. Create a ConfigMap with the script
+
+    ```sh
+    kubectl create configmap hello-world-rust --from-file ./tests/hello-world/script.js
+    ```
+
+3. Run the test by applying the `TestRun` resource
+
+    ```sh
+    kubectl apply -f tests/hello-world/test-run.yaml 
+    ```
+
+4. Once the `hello-world-rust-1-*` Pod has `Completed` you can get the results of the test:
+
+    ```sh
+    kubectl logs hello-world-rust-1-*
+    ```
+    
+    Output should look similar to:
+
+    ```sh
+         ✓ response code was 200
+     ✓ body message was 'Hello, World'
+
+     █ setup
+
+     checks.........................: 100.00% ✓ 39054      ✗ 0
+     data_received..................: 2.6 MB  129 kB/s
+     data_sent......................: 2.2 MB  110 kB/s
+     http_req_blocked...............: avg=19.56µs  min=1.89µs   med=8.94µs   max=37.92ms p(90)=11.68µs  p(95)=12.8µs
+     http_req_connecting............: avg=8.27µs   min=0s       med=0s       max=37.71ms p(90)=0s       p(95)=0s
+     http_req_duration..............: avg=2.46ms   min=301.71µs med=1.95ms   max=38.38ms p(90)=3.8ms    p(95)=4.98ms
+       { expected_response:true }...: avg=2.46ms   min=301.71µs med=1.95ms   max=38.38ms p(90)=3.8ms    p(95)=4.98ms
+     http_req_failed................: 0.00%   ✓ 0          ✗ 19527
+     http_req_receiving.............: avg=116.57µs min=19.83µs  med=108.82µs max=6.14ms  p(90)=153.71µs p(95)=186.96µs
+     http_req_sending...............: avg=41.5µs   min=6.31µs   med=36.68µs  max=1.75ms  p(90)=52.72µs  p(95)=67.84µs
+     http_req_tls_handshaking.......: avg=0s       min=0s       med=0s       max=0s      p(90)=0s       p(95)=0s
+     http_req_waiting...............: avg=2.3ms    min=266.22µs med=1.79ms   max=38.2ms  p(90)=3.62ms   p(95)=4.77ms
+     http_reqs......................: 19527   973.225586/s
+     iteration_duration.............: avg=103.14ms min=38.15µs  med=102.67ms max=145.6ms p(90)=104.54ms p(95)=105.78ms
+     iterations.....................: 19527   973.225586/s
+     vus............................: 19      min=0        max=200
+     vus_max........................: 200     min=200      max=200
+     ```
+
+## Executing a Test with Automatic Application Configuration
+
+Use the [`run.sh`](./tests/with-k8s-extension/run.sh) script to execute a test that also configures the applications. For now, the script assumes a single node cluster. Specify the node IP address in `NODE_IP`.
 
 ```sh
  ./tests/run.sh hello-world $NODE_IP datadog
@@ -46,7 +101,7 @@ Use the [`run.sh`](./tests/run.sh) to execute a test. For now, the script assume
 
 This will start the K6 test, which will create a `SpinApp` and an `Ingress` in your Kubernetes cluster and perform the tests defined in the script. At the end of the test, it will delete the created Kubernetes resources. Test results should be visible in Datadog.
 
-## Creating a Test Configuration File
+### Creating a Test Configuration File
 
 The test configuration file is a JSON file that contains an array of objects. Each object represents a test case and has the following properties:
