+++ tests/hello-world/test-config.json
@@ -1,22 +1,30 @@
 [
     {
-        "name": "hello-world-rust",
+        "service": "hello-world-rust",
         "language": "rust",
-        "image": "ghcr.io/kate-goldenring/test-apps/hello-world-rust:2.3.1"
+        "image": "ghcr.io/kate-goldenring/test-apps/hello-world-rust:2.3.1",
+        "route": "hello",
+        "port": 80
     },
     {
-        "name": "hello-world-go",
+        "service": "hello-world-go",
         "language": "go",
-        "image": "ghcr.io/kate-goldenring/test-apps/hello-world-go:2.3.1"
+        "image": "ghcr.io/kate-goldenring/test-apps/hello-world-go:2.3.1",
+        "route": "hello",
+        "port": 81
     },
     {
-        "name": "hello-world-js",
+        "service": "hello-world-js",
         "language": "js",
-        "image": "ghcr.io/kate-goldenring/test-apps/hello-world-js:2.3.1"
+        "image": "ghcr.io/kate-goldenring/test-apps/hello-world-js:2.3.1",
+        "route": "hello",
+        "port": 82
     },
     {
-        "name": "hello-world-py",
+        "service": "hello-world-py",
         "language": "py",
-        "image": "ghcr.io/kate-goldenring/test-apps/hello-world-py:2.3.1"
+        "image": "ghcr.io/kate-goldenring/test-apps/hello-world-py:2.3.1",
+        "route": "hello", 
+        "port": 83
     }
 ]
