+++ tests/with-k8s-extension/hello-world/options.json
@@ -1,0 +1,12 @@
+{
+  "stages": [
+    {
+      "duration": "30s",
+      "target": 500
+    }
+  ],
+  "noConnectionReuse": true,
+  "tags": {
+    "scenario": "hello-world"
+  }
+}
