+++ tests/with-k8s-extension/run.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-TEST=${TEST:-"hello-world"}
-NODE_IP=${NODE_IP:-"localhost"}
-OUTPUT=${OUTPUT:-"datadog"}
+TEST=${1:-"hello-world"}
+NODE_IP=${2:-"localhost"}
+OUTPUT=${3:-"datadog"}
+EXECUTOR=${4:-"containerd-shim-spin"}
+NO_TEARDOWN=${5:-""}
 # Navigate to the directory containing the script
 path="$(dirname "$0")/$TEST"
 echo "path is $path"
@@ -17,10 +19,8 @@ for entry in $(echo "$test_config_json" | jq -r '.[] | @base64'); do
     }
 
     # Extract values
-    route=$(_jq '.route')
+    ID=$(_jq '.name')
     language=$(_jq '.language')
-    port=$(_jq '.port')
-    endpoint=http://$NODE_IP:$port/$route
 
     # Run command with the appropriate output
     if [ "$OUTPUT" == "prometheus" ];
@@ -30,17 +30,29 @@ for entry in $(echo "$test_config_json" | jq -r '.[] | @base64'); do
         K6_PROMETHEUS_RW_SERVER_URL=http://localhost:9090/api/v1/write \
         K6_PROMETHEUS_RW_TREND_AS_NATIVE_HISTOGRAM=true \
         K6_PROMETHEUS_RW_TREND_STATS='p(99),p(50),min,max,avg' \
-        ./k6 run k6 run --out experimental-prometheus-rw --config $path/options.json --tag language=$language -e ENDPOINT=$endpoint $path/script.js
+        EXECUTOR=$EXECUTOR \
+        TESTCASE_NAME=$ID \
+        TEST_CONFIG_PATH=test-config.json \
+        NODE_IP=$NODE_IP \
+        ./k6 run -o experimental-prometheus-rw --config $path/options.json --tag language=$language $path/script.js $NO_TEARDOWN
     elif [ "$OUTPUT" == "datadog" ];
     then
         echo "Running with Datadog output"
         statsd_addr=$NODE_IP:8125
         K6_STATSD_ADDR=$statsd_addr \
         K6_STATSD_ENABLE_TAGS=true \
-        ./k6 run --out output-statsd --config $path/options.json --tag language=$language -e ENDPOINT=$endpoint $path/script.js
+        EXECUTOR=$EXECUTOR \
+        TESTCASE_NAME=$ID \
+        TEST_CONFIG_PATH=test-config.json \
+        NODE_IP=$NODE_IP \
+        ./k6 run --out output-statsd --config $path/options.json --tag language=$language $path/script.js $NO_TEARDOWN
     else
         echo "Running with no output"
-        ./k6 run --config $path/options.json --tag language=$language $path/script.js -e ENDPOINT=$endpoint $NO_TEARDOWN
+        EXECUTOR=$EXECUTOR \
+        TESTCASE_NAME=$ID \
+        TEST_CONFIG_PATH=test-config.json \
+        NODE_IP=$NODE_IP \
+        ./k6 run --config $path/options.json --tag language=$language $path/script.js $NO_TEARDOWN
     fi
 
     # Sleep between iterations
